tests/requirements.txt
grpcio==1.64.1
protobuf==5.29.3
grpcio-tools==1.64.1